+++ mod-source-record-manager-server/src/main/java/org/folio/services/afterprocessing/AdditionalFieldsUtil.java
@@ -9,6 +9,7 @@ import org.marc4j.MarcJsonWriter;
 import org.marc4j.MarcReader;
 import org.marc4j.MarcStreamWriter;
 import org.marc4j.MarcWriter;
+import org.marc4j.marc.ControlField;
 import org.marc4j.marc.DataField;
 import org.marc4j.marc.MarcFactory;
 import org.marc4j.marc.VariableField;
@@ -42,11 +43,10 @@ public final class AdditionalFieldsUtil {
    */
   public static boolean addFieldToMarcRecord(Record record, String field, char subfield, String value) {
     boolean result = false;
-    try {
+    try (ByteArrayOutputStream os = new ByteArrayOutputStream()) {
       if (record != null && record.getParsedRecord() != null && record.getParsedRecord().getContent() != null) {
-        MarcReader reader = new MarcJsonReader(new ByteArrayInputStream(record.getParsedRecord().getContent().toString().getBytes(StandardCharsets.UTF_8)));
+        MarcReader reader = buildMarcReader(record);
         MarcWriter streamWriter = new MarcStreamWriter(new ByteArrayOutputStream());
-        ByteArrayOutputStream os = new ByteArrayOutputStream();
         MarcJsonWriter jsonWriter = new MarcJsonWriter(os);
         MarcFactory factory = MarcFactory.newInstance();
         if (reader.hasNext()) {
@@ -75,6 +75,76 @@ public final class AdditionalFieldsUtil {
       LOGGER.error("Failed to add additional subfield {} for field {} to record {}", e, subfield, field, record.getId());
     }
     return result;
+  }
+
+  /**
+   * Adds new controlled field to marc record
+   *
+   * @param record record that needs to be updated
+   * @param field  tag of controlled field
+   * @param value  value of the field to add
+   * @return true if succeeded, false otherwise
+   */
+  public static boolean addControlledFieldToMarcRecord(Record record, String field, String value) {
+    boolean result = false;
+    try (ByteArrayOutputStream os = new ByteArrayOutputStream()) {
+      if (record != null && record.getParsedRecord() != null && record.getParsedRecord().getContent() != null) {
+        MarcReader reader = buildMarcReader(record);
+        MarcWriter streamWriter = new MarcStreamWriter(new ByteArrayOutputStream());
+        MarcJsonWriter jsonWriter = new MarcJsonWriter(os);
+        MarcFactory factory = MarcFactory.newInstance();
+        if (reader.hasNext()) {
+          org.marc4j.marc.Record marcRecord = reader.next();
+          ControlField dataField = factory.newControlField(field, value);
+          marcRecord.addVariableField(dataField);
+          // use stream writer to recalculate leader
+          streamWriter.write(marcRecord);
+          jsonWriter.write(marcRecord);
+          record.setParsedRecord(record.getParsedRecord().withContent(new JsonObject(new String(os.toByteArray())).encode()));
+          result = true;
+        }
+      }
+    } catch (Exception e) {
+      LOGGER.error("Failed to add additional controlled field {) to record {}", e, field, record.getId());
+    }
+    return result;
+  }
+
+  /**
+   * remove field from marc record
+   *
+   * @param record record that needs to be updated
+   * @param field  tag of the field
+   * @return true if succeeded, false otherwise
+   */
+  public static boolean removeField(Record record, String field) {
+    boolean result = false;
+    try (ByteArrayOutputStream baos = new ByteArrayOutputStream()) {
+      if (record != null && record.getParsedRecord() != null && record.getParsedRecord().getContent() != null) {
+        MarcReader reader = buildMarcReader(record);
+        MarcWriter marcStreamWriter = new MarcStreamWriter(new ByteArrayOutputStream());
+        MarcJsonWriter marcJsonWriter = new MarcJsonWriter(baos);
+        if (reader.hasNext()) {
+          org.marc4j.marc.Record marcRecord = reader.next();
+          VariableField variableField = marcRecord.getVariableField(field);
+          if (variableField != null) {
+            marcRecord.removeVariableField(variableField);
+          }
+          // use stream writer to recalculate leader
+          marcStreamWriter.write(marcRecord);
+          marcJsonWriter.write(marcRecord);
+          record.setParsedRecord(record.getParsedRecord().withContent(new JsonObject(new String(baos.toByteArray())).encode()));
+          result = true;
+        }
+      }
+    } catch (Exception e) {
+      LOGGER.error("Failed to remove controlled field {) from record {}", e, field, record.getId());
+    }
+    return result;
+  }
+
+  private static MarcReader buildMarcReader(Record record) {
+    return new MarcJsonReader(new ByteArrayInputStream(record.getParsedRecord().getContent().toString().getBytes(StandardCharsets.UTF_8)));
   }
 
   private static VariableField getSingleFieldByIndicators(List<VariableField> list, char ind1, char ind2) {